+++ CHANGELOG.md
@@ -12,7 +12,7 @@
     - Implemented the new ``get_foreign_columns``, ``has_foreign_columns``, and ``get_linked_tables`` also for
       ``AlignedDynamicTable``
     - Added new module ``hdmf.common.hierarchicaltable`` with helper functions to facilitate conversion of
-      hiearchically nested ``DynamicTable`` objects via the following new functions:
+      hierarchically nested ``DynamicTable`` objects via the following new functions:
       - ``to_hierarchical_dataframe`` to merge linked tables into a single consolidated pandas DataFrame.
       - ``drop_id_columns`` to remove "id" columns from a DataFrame.
       - ``flatten_column_index`` to replace a ``pandas.MultiIndex`` with a regular ``pandas.Index``
@@ -24,6 +24,9 @@
 ### Minor improvements
 - Updated ``DynamicTable.to_dataframe()`` and ``DynamicTable.get`` functions to set the ``.name`` attribute
   on generated pandas DataFrame objects. @oruebel (#645)
+- Added ``AlignedDynamicTable.get_colnames(...)`` to support look-up of the full list of columns as the
+  ``AlignedDynamicTable.colnames`` property only includes the columns of the main table for compliance with
+  ``DynamicTable`` @oruebel (#645)
 - Fix documentation for `DynamicTable.get` and `DynamicTableRegion.get`. @rly (#650)
 
 ## HDMF 3.0.1 (July 7, 2021)